+++ test/absinthe/schema/notation/experimental/sdl_extensions_test.exs
@@ -10,7 +10,7 @@ defmodule Absinthe.Schema.Notation.Experimental.SdlExtensionsTest do
 
     directive :feature do
       arg :name, :string
-      on [:scalar]
+      on [:scalar, :schema]
 
       expand(fn _args, node ->
         %{node | __private__: [feature: true]}
@@ -36,6 +36,10 @@ defmodule Absinthe.Schema.Notation.Experimental.SdlExtensionsTest do
       width: Int
     }
 
+    type MyMutationRootType {
+      name: String
+    }
+
     enum Direction {
       NORTH
     }
@@ -54,6 +58,10 @@ defmodule Absinthe.Schema.Notation.Experimental.SdlExtensionsTest do
 
     interface ValuedEntity {
       value: Int
+    }
+
+    extend schema @feature {
+      mutation: MyMutationRootType
     }
 
     extend enum Direction {
@@ -91,6 +99,15 @@ defmodule Absinthe.Schema.Notation.Experimental.SdlExtensionsTest do
     end
 
     def valued_entity_resolve_type(_, _), do: :photo
+  end
+
+  test "can extend schema" do
+    schema_declaration = ExtendedSchema.__absinthe_schema_declaration__()
+
+    assert [%{name: "feature"}] = schema_declaration.directives
+
+    assert [%{name: "query"}, %{name: "mutation", type: %{name: "MyMutationRootType"}}] =
+             schema_declaration.field_definitions
   end
 
   test "can extend enums" do