+++ test/absinthe/schema/type_system_directive_test.exs
@@ -100,6 +100,11 @@ defmodule Absinthe.Schema.TypeSystemDirectiveTest do
 
     @prototype_schema WithTypeSystemDirective
 
+    schema do
+      directive :feature, name: ":schema"
+      field :query, :query
+    end
+
     query do
       field :post, :post do
         directive :feature, name: ":field_definition"
@@ -179,8 +184,7 @@ defmodule Absinthe.Schema.TypeSystemDirectiveTest do
   end
 
   @macro_schema_sdl """
-  "Represents a schema"
-  schema {
+  schema @feature(name: ":schema") {
     query: RootQueryType
   }