+++ test/absinthe/schema_test.exs
@@ -206,6 +206,32 @@ defmodule Absinthe.SchemaTest do
     end
   end
 
+  defmodule RootsSchemaDeclaration do
+    use Absinthe.Schema
+
+    schema do
+      description "Custom schema declaration"
+      field :query, :query
+      field :subscription, :subscription
+    end
+
+    query do
+      field :name,
+        type: :string,
+        args: [
+          family_name: [type: :boolean]
+        ]
+    end
+
+    mutation name: "MyRootMutation" do
+      field :name, :string
+    end
+
+    subscription name: "RootSubscriptionTypeThing" do
+      field :name, :string
+    end
+  end
+
   describe "referenced_types" do
     test "does not contain introspection types" do
       assert !Enum.any?(
@@ -279,6 +305,36 @@ defmodule Absinthe.SchemaTest do
     end
   end
 
+  describe "root fields with custom declaration" do
+    test "custom description" do
+      assert "Custom schema declaration" =
+               Schema.schema_declaration(RootsSchemaDeclaration).description
+    end
+
+    test "it skips the mutation type" do
+      assert [%{name: "subscription"}, %{name: "query"}] =
+               Schema.schema_declaration(RootsSchemaDeclaration).field_definitions
+    end
+
+    test "macro declaration sdl" do
+      assert """
+             "Custom schema declaration"
+             schema {
+               subscription: RootSubscriptionTypeThing
+               query: RootQueryType
+             }
+
+             type RootSubscriptionTypeThing {
+               name: String
+             }
+
+             type RootQueryType {
+               name(familyName: Boolean): String
+             }
+             """ == Schema.to_sdl(RootsSchemaDeclaration)
+    end
+  end
+
   describe "fields" do
     test "have the correct structure in query" do
       assert %Type.Field{name: "name"} = Schema.lookup_type(RootsSchema, :query).fields.name
@@ -300,7 +356,7 @@ defmodule Absinthe.SchemaTest do
   describe "to_sdl/1" do
     test "return schema sdl" do
       assert Schema.to_sdl(SourceSchema) == """
-             \"Represents a schema\"\nschema {\n  query: RootQueryType\n}\n\ntype Foo {\n  name: String\n}\n\n\"can describe query\"\ntype RootQueryType {\n  foo: Foo\n}
+             schema {\n  query: RootQueryType\n}\n\ntype Foo {\n  name: String\n}\n\n\"can describe query\"\ntype RootQueryType {\n  foo: Foo\n}
              """
     end
   end